Solution to Input Two Integers and Print Their Sum

+++ Input Two Integers and Print Their Sum@@6788d5012db25efec946f93d/code.c
@@ -2,7 +2,7 @@
 
 
 int main() {
-    int a,b;
+    double a,b;
     scanf("%d",a);
     scanf("%d",b);
     printf("%d", a+b);
